# ed/l/go/go.gorilla/examples/totd/app/swagger/apidocs/api.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for dispatch lane and flow rack.\nSchemes:\nhttp\nhttps",
    "title": "API.",
    "version": "1.0.0"
  },
  "host": "td.to-api.com",
  "basePath": "/v2",
  "paths": {
    "/clients/{clientId}/mfcs/{mfcId}/configs": {
      "get": {
        "summary": "Get configs.",
        "operationId": "getConfigs",
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/mfcId"
          },
          {
            "type": "string",
            "name": "ClientID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "MfcID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
              "$ref": "#/definitions/listConfigsResponse"
            }
          },
          "400": {
            "description": "Bad Request.",
            "schema": {
              "$ref": "#/definitions/Response400"
            }
          }
        }
      },
      "put": {
        "summary": "Update config.",
        "operationId": "updateConfig",
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/mfcId"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateConfigsRequest"
            }
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateConfigsRequest"
            }
          },
          {
            "type": "string",
            "name": "ClientID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "MfcID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
              "$ref": "#/definitions/updateConfigsResponse"
            }
          },
          "400": {
            "description": "Bad Request.",
            "schema": {
              "$ref": "#/definitions/Response400"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Config": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "example": "maf"
        },
        "count": {
          "type": "integer",
          "example": 17
        },
        "depth": {
          "type": "integer",
          "example": 3
        },
        "env": {
          "type": "string",
          "example": "dev"
        },
        "errorRamp": {
          "type": "integer",
          "example": 99
        },
        "expressLaneMapping": {
          "description": "Express lanes mapping, where key is priority ID and value is actual lane ID.",
          "type": "object",
          "example": "{\"1\":5}"
        },
        "flowRacksMapping": {
          "description": "Flow racks mapping.",
          "type": "object",
          "example": "{\"1\":\"R7L2P01\"}"
        },
        "idGen": {
          "description": "Format string to generate/format id.",
          "type": "string",
          "example": "DISPATCH%02d."
        },
        "laneMapping": {
          "description": "Lanes mapping, where key is priority ID and value is actual lane ID.",
          "type": "object",
          "example": "{\"1\":2, \"2\":1}"
        },
        "mfcId": {
          "type": "string",
          "example": "D01"
        },
        "start": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "Response400": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "error": {
                "description": "Actual error message.",
                "type": "string",
                "example": "Invalid clientId."
              },
              "ref": {
                "description": "In case when provided invalid parameter 'clientId' this field will contain 'clientId' as reference to source of error.",
                "type": "string",
                "example": "clientId"
              },
              "type": {
                "type": "string",
                "example": "validation"
              }
            }
          }
        }
      }
    },
...
    "listConfigsResponse": {
      "type": "object",
      "properties": {
        "configs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Config"
          }
        }
      }
    },
...
    "updateConfigsRequest": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/Config"
        }
      }
    },
    "updateConfigsResponse": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/Config"
        }
      }
    },
...
  },
  "parameters": {
    "clientId": {
      "type": "string",
      "name": "clientId",
      "in": "path",
      "required": true
    },
...
    "mfcId": {
      "type": "string",
      "name": "mfcId",
      "in": "path",
      "required": true
    },
...
  },
...
}